Framework: Extract common from boot and introduce rudimentary API
